Initial commit
Made on the last day of 2018, this is a new repository featuring code
developed over the last four years in various phases.

This code forms the basic SRS model, with functions to sort, search
and select site records, view records and updates, create and edit
update and new site records, and minimal functions for handling site
record lists.

This repo will form the basic deployable codebase, and will be the
central repository from which various descrete development efforts can
be branched.